Handle failure to load ExtendBlock
A block loader might fail loading an ExtendBlock as something it requires
is not present/has been changed (e.g. an itemframe), rather then constantly
throwing stacks record that a block has been lost along with the data for
that block (in future it might be able to regenerate an ExtendBlock from this
information.